add dynamic constant penalty

+++ loss.py
@@ -37,11 +37,11 @@ def dynamic_constant_penalty(outputs, targets, constant_loss):
     direction_loss = [[], [], []]
     for i in range(BATCH_SIZE):
         for j in range(0, 4, 3):
-            direction_loss[0].append(np.array(torch.nn.MSELoss()(outputs[i][j], targets[i][j]), dtype=np.double))
+            direction_loss[0].append(np.array(torch.nn.MSELoss()(outputs[i][j], targets[i][j])))
             for k in range(1, 3):
-                direction_loss[1].append(np.array(torch.nn.MSELoss()(outputs[i][j + k], targets[i][j + k]), dtype=np.double))
+                direction_loss[1].append(np.array(torch.nn.MSELoss()(outputs[i][j + k], targets[i][j + k])))
         for j in range(6, 9):
-            direction_loss[2].append(np.array(torch.nn.MSELoss()(outputs[i][j], targets[i][j]), dtype=np.double))
+            direction_loss[2].append(np.array(torch.nn.MSELoss()(outputs[i][j], targets[i][j])))
 
     dir_percentage = []
     for i in range(2):
@@ -89,8 +89,8 @@ class BCMSELoss(torch.nn.Module):
             extra_scalars.append(extra_scalar)
 
         constant_penalty = [ectra_angles, extra_scalars]
-        constant_penalties = np.array([.0, .0], dtype=np.double)
-        constant_loss = np.array([.0, .0], dtype=np.double)
+        constant_penalties = np.array([.0, .0]).astype(np.double)
+        constant_loss = np.array([.0, .0]).astype(np.double)
         for i in range(2):
             constant = np.array(constant_penalty[i]).sum()
             constant_penalties[i] = constant / BATCH_SIZE / CONSTANT_WEIGHT
